requirements.txt
celery[redis]
Pillow
redis
Flask
Jinja2